python-ssh/requirements.txt
netmiko==2.0.0
pyyaml==3.12
netdev==0.7.1